+++ src/main/java/com/mp/cinepop/account/controller/AccountController.java
@@ -1,86 +1,50 @@
 package com.mp.cinepop.account.controller;
 
-import java.security.NoSuchAlgorithmException;
-
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.mp.cinepop.account.model.AccountService;
 import com.mp.cinepop.account.model.AccountVO;
-import com.mp.cinepop.common.HashingUtil;
-import com.mp.cinepop.hash.model.hashVO;
+import com.mp.cinepop.qna.controller.QnaController;
+import com.mp.cinepop.qna.model.QnaVO;
 
 @Controller
 @RequestMapping("/")
 public class AccountController {
-	private static final Logger logger
-	=LoggerFactory.getLogger(AccountController.class);
+	private static final Logger logger = LoggerFactory.getLogger(QnaController.class);
 	
 	private final AccountService accountService;
-	private final HashingUtil hash;
-	
-	@Autowired
-	public AccountController(AccountService accountService, HashingUtil hash) {
-		this.accountService = accountService;
-		this.hash = hash;
-	}
-	
-	
-	@GetMapping("register/register")
-	public void register() {
-		logger.info("회원가입화면");
-	}
-	
 
-	@PostMapping("register/register")
-	public String register_post(@ModelAttribute AccountVO vo, @ModelAttribute hashVO hashvo,@RequestParam String email2,String email3,String pwd, String yy, String MM, String dd, Model model) throws NoSuchAlgorithmException {
-		logger.info("회원가입 처리 , 파라미터 vo={}",vo);
-		
-	
-		 if(vo.getId()==null || vo.getId().isEmpty()) { 
-			 email2=""; email3=""; 
-		 }else if(email3!=null && !email3.isEmpty()) {
-			 email2=email3; }
-		 
-		int cnt=accountService.insertAccount(vo);
-		
-		vo.setId(vo.getId()+"@"+ email2);
-		
-		String salt=hash.makeNewSalt();
-		hashvo.setSalt(salt);
-		
-		String digest =hash.hashing(pwd, salt);
-		hashvo.setDigest(digest);
-		logger.info("pwd : ",pwd);
-		
-		cnt=accountService.insertHash(hashvo);
-		logger.info("회원가입 결과,cnt={}",cnt);
-	
-		return "redirect:/login/login";
-	} 
-	
-	@GetMapping("mypage/withdrawal")
-	public void withdrawal_get() {
-		logger.info("회원탈퇴 페이지");
+	@Autowired
+	public AccountController(AccountService accountService) {
+		this.accountService = accountService;
+		logger.info("AccountController생성자 주입!!");
 	}
 	
 	/*
-	 * @PostMapping("mypage/withdrawal") public String withdrawal_post(@RequestParam
-	 * String pwd, HttpSession session,HttpServletResponse response, Model model) {
-	 * String id = (String)session.getAttribute("id");
-	 * logger.info("회원탈퇴 처리, 파라미터 id={},pwd={}",id,pwd);
+	 * @RequestMapping("qna/qna_detail") // 새 값 넣기 public String
+	 * detail(@RequestParam(defaultValue = "0") int aNo, Model model) {
+	 * logger.info("회원정보 상세보기 파라미터 no={}", aNo);
 	 * 
-	 * }
+	 * if(aNo==0) { model.addAttribute("msg", "잘못된 url입니다.");
+	 * model.addAttribute("url", "/qna/qna_list"); // 새 값 넣기
+	 * 
+	 * return "/qna/qna_list"; // 새 값 넣기 }
+	 * 
+	 * AccountSEVO vo = accountService.selectByNo(aNo);
+	 * logger.info("회원정보 상세보기 결과 vo={}", vo);
+	 * 
+	 * model.addAttribute("vo", vo);
+	 * 
+	 * return "qna/qna_detail"; // 새 값 넣기 }
 	 */
+	
+
+	
+	
 }
