+++ src/main/java/config/mybatis/mapper/oracle/account.xml
@@ -3,19 +3,6 @@
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="com.mp.cinepop.account.model.AccountDAO">
-	<insert id="insertAccount" parameterType="accountVo">
-		<selectKey keyProperty="aNo" resultType="int" order="BEFORE">
-			select account_seq.nextval from dual
-		</selectKey>
-		insert into account(aNo, aName,id,postcode1,address,detailAddress1,tel,pwd)
-		values(#{aNo}, #{aName},#{id},#{postcode1},#{address},#{detailAddress1},#{tel},#{pwd})
-	</insert>
-	
-	<insert id="insertHash" parameterType="hashVo">
-		insert into hash(id,salt,digest)
-		values(#{id},#{salt},#{digest})
-	</insert>
-	
 	<select id="selectPwd" parameterType="string" resultType="string">
 		select pwd from account
 		where id=#{id}
