+++ src/main/java/config/mybatis/mapper/oracle/accountInsert.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
+"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.mp.cinepop.accountInsert.model.AccountInsertDAO">
+	<insert id="insertAccount" parameterType="accountInsertVo">
+		<selectKey keyProperty="aNo" resultType="int" order="BEFORE">
+			select account_seq.nextval from dual
+		</selectKey>
+		insert into account(aNo, aName,id,postcode1,address,detailAddress1,tel,pwd)
+		values(#{aNo}, #{aName},#{id},#{postcode1},#{address},#{detailAddress1},#{tel},#{pwd})
+	</insert>
+	
+	<insert id="insertHash" parameterType="hashVo">
+		insert into hash(id,salt,digest)
+		values(#{id},#{salt},#{digest})
+	</insert>
+	
+	<select id="selectPwd" parameterType="string" resultType="string">
+		select pwd from account
+		where id=#{id}
+	</select>
+	
+	<select id="selectByUserid" parameterType="string" resultType="accountInsertVo">
+		select * from account
+		where id=#{id}
+	</select>
+	
+	<select id="hashCheck" parameterType="string" resultType="hashVo">
+		select * from hash
+		where id=#{id}
+	</select>
+	
+	<select id="selectDigest" parameterType="string" resultType="string">
+		select digest from hash
+		where id=#{id}
+	</select>
+</mapper>
